feat: cleanup old GWASUI app code
and keep only V2 and the Results viewer app
