+++ src/Analysis/GWASResults/constants.js
@@ -1,0 +1,7 @@
+export const gwasStatus = {
+  pending: 'Pending',
+  running: 'Running',
+  succeeded: 'Succeeded',
+  failed: 'Failed',
+  error: 'Error',
+};
