--- src/Analysis/GWASUIApp/ReduxGWASUIApp.js
@@ -1,9 +1,0 @@
-import { connect } from 'react-redux';
-import GWASUIApp from './GWASUIApp';
-
-const mapStateToProps = (state) => ({
-  userAuthMapping: state.userAuthMapping,
-});
-
-const ReduxGWASUIApp = connect(mapStateToProps)(GWASUIApp);
-export default ReduxGWASUIApp;
