+++ src/Analysis/GWASV2/Components/SearchBar/SearchBar.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { SearchOutlined } from '@ant-design/icons';
-import '../../../GWASUIApp/GWASUIApp.css';
+import '../../../GWASResults/GWASUIApp.css';
 
 const SearchBar = ({ searchTerm, handleSearch, field = 'variable name' }) => (
   <div data-tour='search-bar' className='GWASUI-searchContainer'>
