+++ src/Analysis/GWASV2/Components/SelectCohort/SelectCohort.stories.jsx
@@ -8,7 +8,7 @@ import { SourceContextProvider } from '../../Utils/Source';
 import ACTIONS from '../../Utils/StateManagement/Actions';
 import './SelectCohort.css';
 import '../../../GWASV2/GWASV2.css';
-import '../../../GWASUIApp/GWASUIApp.css';
+import '../../../GWASResults/GWASUIApp.css';
 
 export default {
   title: 'Tests3/GWASV2/SelectCohort/SelectCohort',
