+++ src/Analysis/GWASV2/Components/SelectCohort/SelectCohortDropDown.stories.jsx
@@ -6,7 +6,7 @@ import { Space } from 'antd';
 import { SourceContextProvider } from '../../Utils/Source';
 import './SelectCohort.css';
 import '../../../GWASV2/GWASV2.css';
-import '../../../GWASUIApp/GWASUIApp.css';
+import '../../../GWASResults/GWASUIApp.css';
 
 export default {
   title: 'Tests3/GWASV2/SelectCohort/SelectCohortDropDown',
